requirements.txt
requests>=2.28.0
tqdm>=4.65.0
numpy>=1.24.0
ccxt>=4.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0
flask>=3.0.0
python-dotenv>=1.0.0
fredapi>=0.5.0
scikit-learn>=1.0.0
statsmodels>=0.14.0
